Registration and user confirmation added.

+++ src/main/java/edu/ben/util/Email.java
@@ -11,7 +11,7 @@ public class Email {
 
         String code = "";
         for (int i = 0; i < 6; i++) {
-            code += Math.random() * 10;
+            code += (int)(Math.random() * 10);
         }
 
         sendEmail("Your verification code is: " + code,
